+++ server/src/utils/selector-analyzer.ts
